Added compatibility to OS X, fixed freebayes bug
dDocent.FB is now more compatible with default installations of OS X.

More importantly, this version of dDocent has some changes to the formation of reference contigs.  The reference contigs now have a single N placed on both 5' and 3' ends of the contig.  The current version of FreeBayes was ignoring some reads that mapped exactly to the 5' end of a reference contig, and adding the N padding fixes this issue.  

Also, dDocent now outputs the reference in the correct orientation.  Previous versions output reverse complemented reference contigs. 

It is recommended that all previous analyses are rerun with the current version.