Major Update to Version 1.2
This update brings some major changes to dDocent.FB.  This version now checks that all necessary dependencies are installed and will not run unless they are.  It also is more flexible in finding where necessary files and executables are.  However, they still need to be in the $PATH to operate properly.  This version also has upgraded the parallelization of FreeBayes variant calling and now requires GNU-parallel to work properly.  Additional changes include important changes to the default assembly parameters that lead to more accurate assemblies and there is now a filter to remove contigs that are in the 0.01% of coverage.  The extreme coverage contigs are usually either some form of contamination or multi copy loci that would be excluded anyway.  This also prevents hangs in variant calling due to extreme depth.